+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 0.5.1 - 2020-03-15
+
+- ADDED
+- REMOVED
+- IMPROVED
+- BUGFIXES
+  - Tab bar incorrectly sized when resize event is triggered
+  - Tab slide was enabling autoplay by default
+  - Fix btn hoverable & light-hoverable
+  - Caroulix `updateHeight` method rework
+  - Caroulix resize event
+  - Caroulix optimization
+  - Collapsible height in sidenav
+  - Fab hover method
+
 # 0.5.0 - 2020-03-04
 
 - ADDED